// client/src/shared/shared.interface.ts
export interface IGigCardItemModal {
    overlay: boolean;
    deleteApproval: boolean;
}